requirements.txt
fastapi
uvicorn
playwright
asyncpg
python-dotenv
passlib
bcrypt
pyjwt